# Conflicts:
#	src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
#	src/test/java/com/catalyst/collector/daos/hibernate/CollectionsDaoHibernateTest.java
#	src/test/java/com/catalyst/collector/services/impl/CollectionsServiceImplTest.java
finished feature

+++ src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
@@ -8,13 +8,6 @@ import com.catalyst.collector.entities.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.catalyst.collector.daos.CollectionsDao;
-import com.catalyst.collector.entities.Color;
-import com.catalyst.collector.entities.Category;
-import com.catalyst.collector.entities.Keyword;
-import com.catalyst.collector.entities.Collectible;
-
-import com.catalyst.collector.entities.Age;
-
 import com.catalyst.collector.services.CollectionsService;
 
 @Service
@@ -27,32 +20,26 @@ public class CollectionsServiceImpl implements CollectionsService {
 		this.collectionsDao = collectionsDao;
 	}
 
-
-
 	@Override
 	public ArrayList<Age> getAgeTypes(){
 		return collectionsDao.getAgeTypes();
 	}
 
 	@Override
-	public boolean addAge(Age age) {
+	public void addAge(Age age) {
 		String ageString = age.getAge();
-		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+		if(ageString != null && ageString.length() <= 255 && !ageString.matches(".*\\d.*") && !ageString.trim().equals("")) { //Maximum of 255 characters for an age, no digits allowed
 			collectionsDao.addAge(age);
-			return true;
 		}
-		return false;
 	}
 
 
 	@Override
-	public boolean updateAge(Age age){
+	public void updateAge(Age age){
 		String ageString = age.getAge();
-		if(ageString.length() <= 255 && !ageString.matches(".*\\d.*")) { //Maximum of 255 characters for an age, no digits allowed
+		if(ageString != null && ageString.length() <= 255 && !ageString.matches(".*\\d.*") && !ageString.trim().equals("")) { //Maximum of 255 characters for an age, no digits allowed
 			collectionsDao.updateAge(age);
-			return true;
 		}
-		return false;
 
 	}
 
@@ -74,7 +61,7 @@ public class CollectionsServiceImpl implements CollectionsService {
 
 	@Override
 	public boolean addCategory(Category category) {
-		if(category.getCategory() == null || ((category.getCategory()).trim()).equals(""))
+		if(category.getCategory() == null || ((category.getCategory()).trim()).equals("")|| category.getCategory().matches(".*\\d.*"))
 		{
 			return false;
 		}
@@ -89,8 +76,9 @@ public class CollectionsServiceImpl implements CollectionsService {
 
 	@Override
 	public boolean updateCategory(int id, Category category) {
+
 		category.setId(id);
-		if(category.getCategory() == null || ((category.getCategory()).trim()).equals(""))
+		if(category.getCategory() == null || ((category.getCategory()).trim()).equals("")  || category.getCategory().matches(".*\\d.*"))
 		{
 			return false;
 		}
@@ -103,7 +91,11 @@ public class CollectionsServiceImpl implements CollectionsService {
 
 	@Override
 	public boolean deleteCategory(int id) {
-		return collectionsDao.deleteCategory(id);
+		if(id > 0)
+		{
+			return collectionsDao.deleteCategory(id);
+		}
+		return false;
 	}
 
 	@Override
@@ -112,12 +104,10 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 	@Override
 	public boolean addColor(Color addedColor) {
-		try{
-			collectionsDao.addColor(addedColor);
-		}
-		catch(Exception e){
-			return false;
-		}
+		 if (addedColor.getColor() == null || addedColor.getColor().trim().equals("") || addedColor.getColor().length() > 255){
+	            return false;
+		 }
+		collectionsDao.addColor(addedColor);
 		return true;
 		}
 	@Override
@@ -132,36 +122,26 @@ public class CollectionsServiceImpl implements CollectionsService {
 	}
 
 	@Override
-	public boolean updateColor(int id, String color){
-		try{
-		Color c = getColor(id);
-		c.setColor(color);
+	public boolean updateColor(Color c){
+		 if (c.getColor() == null || c.getColor().trim().equals("") || c.getColor().length() > 255){
+	            return false;
+		 }
 		collectionsDao.updateColor(c);
-		}
-		catch(Exception e){
-			return false;
-			}
 		return true;
 	}
 	public Color getColor(int id){
-		List<Color> colors = getColorList();
-		for(Color c: colors){
-			if(c.getId() == id){
-				return c;
-			}
-		}
-		return null;
+		return collectionsDao.getColor(id);
 
-	}
-
-	@Override
-	public Color getByColorId(int colorId){
-		return collectionsDao.getColor(colorId);
 	}
 
     @Override
     public ArrayList<Keyword> getAllKeywords() {
         return collectionsDao.getAllKeywords();
+    }
+
+    @Override
+    public ArrayList<Keyword> getKeywordsByLetter(char letter){
+    	return collectionsDao.getKeywordsByLetter(letter);
     }
 
     @Override
